Move the compat library to a sensible location
